pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faceverify-sdk"
version = "1.0.0rc1"
description = "A modular, open-source face verification SDK for Python"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "nayandas69", email = "nayanchandradas@hotmail.com"}
]
maintainers = [
    {name = "nayandas69", email = "nayanchandradas@hotmail.com"}
]
keywords = [
    "face-recognition",
    "face-verification",
    "face-detection",
    "deep-learning",
    "computer-vision",
    "biometrics",
    "identity-verification",
    "facenet",
    "deepface",
    "opencv",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",
    "scikit-learn>=1.0.0",
    "deepface>=0.0.79",
    "tf-keras>=2.15.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tqdm>=4.64.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=1.0.0",
]
gpu = [
    "tensorflow>=2.10.0",
    "onnxruntime-gpu>=1.12.0",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6",
]
detection = [
    "mtcnn>=0.1.1",
    "mediapipe>=0.10.0",
]
all = [
    "faceverify-sdk[dev,docs,api,detection]",
]

[project.urls]
Homepage = "https://github.com/nayandas69/faceverify"
Documentation = "https://github.com/nayandas69/faceverify#readme"
Repository = "https://github.com/nayandas69/faceverify"
Changelog = "https://github.com/nayandas69/faceverify/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/nayandas69/faceverify/issues"

[project.scripts]
faceverify = "faceverify.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
faceverify = ["py.typed", "configs/*.yaml"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/setup.py"]
known_first_party = ["faceverify"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "PIL.*",
    "tensorflow.*",
    "tf_keras.*",
    "onnxruntime.*",
    "sklearn.*",
    "deepface.*",
    "mtcnn.*",
    "mediapipe.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::FutureWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests requiring GPU",
]

[tool.coverage.run]
source = ["src/faceverify"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true
fail_under = 70
